+++ wdl-analysis/src/graph.rs
@@ -26,6 +26,7 @@ use tokio::runtime::Handle;
 use tracing::debug;
 use tracing::info;
 use url::Url;
+use wdl_ast::AstNode;
 use wdl_ast::Diagnostic;
 use wdl_ast::SyntaxNode;
 use wdl_ast::Validator;
@@ -357,7 +358,7 @@ impl DocumentGraphNode {
 
         Ok(ParseState::Parsed {
             version,
-            root: document.syntax().green().into(),
+            root: document.inner().green().into(),
             lines,
             diagnostics,
         })
@@ -422,14 +423,15 @@ pub struct DocumentGraph {
 impl DocumentGraph {
     /// Add a node to the document graph.
     pub fn add_node(&mut self, uri: Url, rooted: bool) -> NodeIndex {
-        let index = if let Some(index) = self.indexes.get(&uri) {
-            *index
-        } else {
-            debug!("inserting `{uri}` into the document graph");
-            let uri = Arc::new(uri);
-            let index = self.inner.add_node(DocumentGraphNode::new(uri.clone()));
-            self.indexes.insert(uri, index);
-            index
+        let index = match self.indexes.get(&uri) {
+            Some(index) => *index,
+            _ => {
+                debug!("inserting `{uri}` into the document graph");
+                let uri = Arc::new(uri);
+                let index = self.inner.add_node(DocumentGraphNode::new(uri.clone()));
+                self.indexes.insert(uri, index);
+                index
+            }
         };
 
         if rooted {