+++ wdl-engine/src/stdlib/prefix.rs
@@ -5,6 +5,7 @@ use wdl_analysis::types::PrimitiveType;
 use wdl_ast::Diagnostic;
 
 use super::CallContext;
+use super::Callback;
 use super::Function;
 use super::Signature;
 use crate::Array;
@@ -51,7 +52,7 @@ pub const fn descriptor() -> Function {
         const {
             &[Signature::new(
                 "(String, Array[P]) -> Array[String] where `P`: any primitive type",
-                prefix,
+                Callback::Sync(prefix),
             )]
         },
     )
@@ -65,10 +66,12 @@ mod test {
     use crate::v1::test::TestEnv;
     use crate::v1::test::eval_v1_expr;
 
-    #[test]
-    fn prefix() {
-        let mut env = TestEnv::default();
-        let value = eval_v1_expr(&mut env, V1::Zero, "prefix('foo', [1, 2, 3])").unwrap();
+    #[tokio::test]
+    async fn prefix() {
+        let env = TestEnv::default();
+        let value = eval_v1_expr(&env, V1::Zero, "prefix('foo', [1, 2, 3])")
+            .await
+            .unwrap();
         let elements: Vec<_> = value
             .as_array()
             .unwrap()
@@ -78,7 +81,9 @@ mod test {
             .collect();
         assert_eq!(elements, ["foo1", "foo2", "foo3"]);
 
-        let value = eval_v1_expr(&mut env, V1::Zero, "prefix('foo', [1.0, 1.1, 1.2])").unwrap();
+        let value = eval_v1_expr(&env, V1::Zero, "prefix('foo', [1.0, 1.1, 1.2])")
+            .await
+            .unwrap();
         let elements: Vec<_> = value
             .as_array()
             .unwrap()
@@ -88,8 +93,9 @@ mod test {
             .collect();
         assert_eq!(elements, ["foo1.000000", "foo1.100000", "foo1.200000"]);
 
-        let value =
-            eval_v1_expr(&mut env, V1::Zero, "prefix('foo', ['bar', 'baz', 'qux'])").unwrap();
+        let value = eval_v1_expr(&env, V1::Zero, "prefix('foo', ['bar', 'baz', 'qux'])")
+            .await
+            .unwrap();
         let elements: Vec<_> = value
             .as_array()
             .unwrap()
@@ -99,7 +105,9 @@ mod test {
             .collect();
         assert_eq!(elements, ["foobar", "foobaz", "fooqux"]);
 
-        let value = eval_v1_expr(&mut env, V1::Zero, "prefix('foo', [1, None, 3])").unwrap();
+        let value = eval_v1_expr(&env, V1::Zero, "prefix('foo', [1, None, 3])")
+            .await
+            .unwrap();
         let elements: Vec<_> = value
             .as_array()
             .unwrap()
@@ -109,7 +117,9 @@ mod test {
             .collect();
         assert_eq!(elements, ["foo1", "foo", "foo3"]);
 
-        let value = eval_v1_expr(&mut env, V1::One, "prefix('foo', [])").unwrap();
+        let value = eval_v1_expr(&env, V1::One, "prefix('foo', [])")
+            .await
+            .unwrap();
         assert!(value.unwrap_array().is_empty());
     }
 }